+++ src/modules/jobs/DamageMapSection.js
@@ -2,13 +2,14 @@ import DamageMap from '../../components/DamageMap';
 
 export default function DamageMapSection({
   canWrite = true,
+  className = '',
   damageMap,
   instrumentType,
   onChange,
   onViewImageUpload
 }) {
   return (
-    <section>
+    <section className={className || undefined}>
       <h3>Damage Map</h3>
       <DamageMap
         instrumentType={instrumentType}